Optionally add `AuthSession` to `redirectUrl`
If `addAuthSessionToRedirectUrl` is `true`, the `AuthSession` cookie
value will be added as `AuthSession` query parameter to the
`redirectUrl` in the response after following the sign-in link. This is
for the case that browsers don't accept third-party cookies, and
requests to the couch_email_auth endpoint are cross-origin. If
`addAuthSessionToRedirectUrl` is `true`, clients can then set the
`AuthSession` cookie themselves.

Closes #43.